mutating iterators, broken branch swapping, but lots of perf improvements

+++ bp_vector/bp_vector_glob.cc
@@ -1,5 +1,5 @@
 
 #include "bp_vector.h"
 
-std::atomic<int16_t> bp_vector_glob::unique_id{1};
+std::atomic<int32_t> bp_vector_glob::unique_id{1};
